+++ ForecastService/Models/OpenMeteo/Responses/Hourly.cs
@@ -7,17 +7,17 @@ namespace ForecastService.Models.OpenMeteo.Responses
     public class Hourly
     {
         public List<string> time { get; set; } = new();
-        public List<double?>? dust { get; set; } = new();
-        public List<double?>? alder_pollen { get; set; } = new();
-        public List<double?>? birch_pollen { get; set; } = new();
-        public List<double?>? grass_pollen { get; set; } = new();
-        public List<double?>? mugwort_pollen { get; set; } = new();
-        public List<double?>? olive_pollen { get; set; } = new();
-        public List<double?>? ragweed_pollen { get; set; } = new();
-        public List<double?>? pm10 { get; set; } = new();
-        public List<double?>? pm2_5 { get; set; } = new();
-        public List<int?>? european_aqi { get; set; } = new();
-        public List<double?>? ozone { get; set; } = new();
-        public List<double?>? nitrogen_dioxide { get; set; } = new();
+        public List<float?>? dust { get; set; } = new();
+        public List<float?>? alder_pollen { get; set; } = new();
+        public List<float?>? birch_pollen { get; set; } = new();
+        public List<float?>? grass_pollen { get; set; } = new();
+        public List<float?>? mugwort_pollen { get; set; } = new();
+        public List<float?>? olive_pollen { get; set; } = new();
+        public List<float?>? ragweed_pollen { get; set; } = new();
+        public List<float?>? pm10 { get; set; } = new();
+        public List<float?>? pm2_5 { get; set; } = new();
+        public List<float?>? european_aqi { get; set; } = new();
+        public List<float?>? ozone { get; set; } = new();
+        public List<float?>? nitrogen_dioxide { get; set; } = new();
     }
 }
