+++ Server/Application/DTOs/PlotHnadling/PlotDto.cs
@@ -7,7 +7,7 @@ namespace API.Application.DTOs.PlotHnadling
     {
         public Guid plotId { get; set; }
         public required Geometry polygon { get; set; }
-        public double area { get; set; }
+        public float area { get; set; }
 
         //agricultural info
         public ICollection<CropType>? cropTypes { get; set; }
